Remove ApiError and ApiResult
